Changed int to float

+++ simpleinterest.py
@@ -1,5 +1,5 @@
 principal = int(input("Enter principal amount : "))
-rate = int(input("Enter rate of interest (per annum): "))
-time = int(input("Enter time period (in years) : "))
+rate = float(input("Enter rate of interest (per annum): "))
+time = float(input("Enter time period (in years) : "))
 si = (principal*rate*time)/100
 print("Simple interest  : ", si)
